server/requirements.txt
fastapi==0.115.6
openai==1.59.6
python-dotenv==1.0.1
uvicorn[standard]==0.34.0
